pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "peer-ai"
version = "0.1.0"
description = "AI-powered code review for GitHub, GitLab, and git servers"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Antonio Stepien", email = "antonio@innovateops.dev" }
]
keywords = ["code-review", "ai", "security", "static-analysis", "github", "gitlab"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Security",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "tree-sitter>=0.21.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "pygithub>=2.1.0",
    "python-gitlab>=4.2.0",
]

[project.optional-dependencies]
train = [
    "datasets>=2.16.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.42.0",
    "accelerate>=0.25.0",
    "trl>=0.7.0",
    "wandb>=0.16.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
all = ["peer-ai[train,dev]"]

[project.scripts]
peer-ai = "peer_ai.cli:main"

[project.urls]
Homepage = "https://github.com/antonioinnovateops/peer-ai"
Documentation = "https://github.com/antonioinnovateops/peer-ai#readme"
Repository = "https://github.com/antonioinnovateops/peer-ai"
Issues = "https://github.com/antonioinnovateops/peer-ai/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/peer_ai"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "C4", "PT"]
ignore = ["S101"]  # Allow assert in tests

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
